+++ src/main/java/atas/logic/Logic.java
@@ -7,9 +7,9 @@ import atas.logic.commands.CommandResult;
 import atas.logic.commands.exceptions.CommandException;
 import atas.logic.parser.exceptions.ParseException;
 import atas.model.Model;
-import atas.model.ReadOnlyStudentList;
 import atas.model.session.Attributes;
 import atas.model.session.Session;
+import atas.model.student.ReadOnlyStudentList;
 import atas.model.student.Student;
 import javafx.collections.ObservableList;
