+++ src/main/java/atas/logic/LogicManager.java
@@ -12,9 +12,9 @@ import atas.logic.commands.exceptions.CommandException;
 import atas.logic.parser.AtasParser;
 import atas.logic.parser.exceptions.ParseException;
 import atas.model.Model;
-import atas.model.ReadOnlyStudentList;
 import atas.model.session.Attributes;
 import atas.model.session.Session;
+import atas.model.student.ReadOnlyStudentList;
 import atas.model.student.Student;
 import atas.storage.Storage;
 import javafx.collections.ObservableList;
