+++ src/main/java/atas/logic/commands/studentlist/ClearStudentListCommand.java
@@ -5,7 +5,7 @@ import static java.util.Objects.requireNonNull;
 import atas.logic.commands.CommandResult;
 import atas.logic.commands.confirmation.DangerousCommand;
 import atas.model.Model;
-import atas.model.StudentList;
+import atas.model.student.StudentList;
 
 /**
  * Clears the student list.
