+++ src/main/java/atas/model/Model.java
@@ -11,6 +11,7 @@ import atas.model.session.IndexRange;
 import atas.model.session.Session;
 import atas.model.session.SessionList;
 import atas.model.session.SessionName;
+import atas.model.student.ReadOnlyStudentList;
 import atas.model.student.Student;
 import javafx.collections.ObservableList;
