+++ src/test/java/atas/logic/LogicManagerTest.java
@@ -32,10 +32,10 @@ import atas.logic.commands.studentlist.ListStudentsCommand;
 import atas.logic.parser.exceptions.ParseException;
 import atas.model.Model;
 import atas.model.ModelManager;
-import atas.model.ReadOnlySessionList;
-import atas.model.ReadOnlyStudentList;
 import atas.model.UserPrefs;
 import atas.model.memo.Memo;
+import atas.model.session.ReadOnlySessionList;
+import atas.model.student.ReadOnlyStudentList;
 import atas.model.student.Student;
 import atas.storage.JsonAtasStorage;
 import atas.storage.JsonSessionListStorage;
