+++ src/test/java/atas/logic/commands/confirmation/ConfirmationAcceptCommandTest.java
@@ -20,9 +20,9 @@ import atas.logic.commands.studentlist.DeleteStudentCommand;
 import atas.logic.commands.studentlist.EditStudentCommand;
 import atas.model.Model;
 import atas.model.ModelManager;
-import atas.model.StudentList;
 import atas.model.UserPrefs;
 import atas.model.student.Student;
+import atas.model.student.StudentList;
 import atas.testutil.EditStudentDescriptorBuilder;
 import atas.testutil.ModelManagerBuilder;
 import atas.testutil.StudentBuilder;
