+++ src/test/java/atas/logic/commands/sessionlist/EditSessionCommandTest.java
@@ -21,9 +21,9 @@ import atas.commons.core.index.Index;
 import atas.logic.commands.studentlist.ClearStudentListCommand;
 import atas.model.Model;
 import atas.model.ModelManager;
-import atas.model.StudentList;
 import atas.model.UserPrefs;
 import atas.model.session.Session;
+import atas.model.student.StudentList;
 import atas.testutil.EditSessionDescriptorBuilder;
 import atas.testutil.ModelManagerBuilder;
 import atas.testutil.SessionBuilder;
