+++ src/test/java/atas/logic/commands/studentlist/ClearStudentListCommandTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 
 import atas.model.Model;
 import atas.model.ModelManager;
-import atas.model.StudentList;
+import atas.model.student.StudentList;
 import atas.testutil.ModelManagerBuilder;
 
 public class ClearStudentListCommandTest {
