+++ src/test/java/atas/logic/commands/studentlist/EditStudentCommandTest.java
@@ -22,9 +22,9 @@ import atas.commons.core.index.Index;
 import atas.logic.commands.studentlist.EditStudentCommand.EditStudentDescriptor;
 import atas.model.Model;
 import atas.model.ModelManager;
-import atas.model.StudentList;
 import atas.model.UserPrefs;
 import atas.model.student.Student;
+import atas.model.student.StudentList;
 import atas.testutil.EditStudentDescriptorBuilder;
 import atas.testutil.ModelManagerBuilder;
 import atas.testutil.StudentBuilder;
