+++ src/test/java/atas/model/session/ParticipationTest.java
@@ -16,8 +16,8 @@ import atas.logic.commands.sessionlist.session.ParticipateCommand;
 import atas.logic.commands.sessionlist.session.PresenceCommand;
 import atas.model.Model;
 import atas.model.ModelManager;
-import atas.model.StudentList;
 import atas.model.UserPrefs;
+import atas.model.student.StudentList;
 import atas.testutil.ModelManagerBuilder;
 
 class ParticipationTest {
