+++ src/test/java/atas/storage/JsonSessionListStorageTest.java
@@ -16,7 +16,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
 import atas.commons.exceptions.DataConversionException;
-import atas.model.ReadOnlySessionList;
+import atas.model.session.ReadOnlySessionList;
 import atas.model.session.SessionList;
 
 public class JsonSessionListStorageTest {
